// src/app/web-api.ts
export class FinanceRecord implements IFinanceRecord {
  amount?: number;
  createdDate?: Date;
  id?: number;
  type?: number;

  constructor(data?: IFinanceRecord) {
    if (data) {
      for (var property in data) {
        if (data.hasOwnProperty(property))
          (<any>this)[property] = (<any>data)[property];
      }
    }
  }

  init(_data?: any) {
    if (_data) {
      this.amount = _data["amount"];
      this.createdDate = _data["createdDate"] ? new Date(_data["createdDate"].toString()) : <any>undefined;
      this.id = _data["id"];
      this.type = _data["type"];
    }
  }

  static fromJS(data: any): FinanceRecord {
    data = typeof data === 'object' ? data : {};
    let result = new FinanceRecord();
    result.init(data);
    return result;
  }

  toJSON(data?: any) {
    data = typeof data === 'object' ? data : {};
    data["amount"] = this.amount;
    data["createdDate"] = this.createdDate ? this.createdDate.toISOString() : <any>undefined;
    data["id"] = this.id;
    data["type"] = this.type;
    return data;
  }
}

export interface IFinanceRecord {
  amount?: number;
  createdDate?: Date;
  id?: number;
  type?: number;
}